pppossh: initial commit.
This patch adds protocol support for PPP over SSH.  The protocol name is
'pppossh' with the following options.

 - server, SSH server name.  Required
 - port, SSH server port.  Defaults to 22.
 - sshuser, SSH login username.  Required.
 - identity, list of client private key files.  ~/.ssh/id_{rsa,dsa} will
   be used if no identity file was specified and at least one of them must
   be valid for the public key authentication to proceed.
 - ipaddr, local ip address to be assigned.
 - peeraddr, peer ip address to be assigned.
 - ssh_options, extra ssh client options.

Because the protocol script file ppp.sh will be called with $HOME set to
'/', we need to explicitly set it to the right value so that dropbear
client can read '~/known_hosts' correctly.