pppossh: simplify INCLUDE_ONLY handling.
If called, it is expected that proto shell scripts will be feeded with
at least an action parameter (dump, setup, teardown, and renew).

If sourced, they are expected to be quiet when [ -n "$INCLUDE_ONLY" ] is
true.  However, if the proto script is not to be sourced (included) by
others, then slack this requirement as does in this commit by following
what's already done in wwan package.